run test cases in multiprocess
